# metools-backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: MeTools
  description: Documentation to MeTools API
  license:
    name: ''
  version: 0.1.0
paths:
  /api/v1/rzd/tasks:
    get:
      tags:
      - tasks
      operationId: list_tasks
      parameters:
      - name: X-API-AUTH-TOKEN
        in: header
        description: Auth token
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListTasks'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - tasks
      operationId: create_task
      parameters:
      - name: X-API-AUTH-TOKEN
        in: header
        description: Auth token
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskData'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCreateTask'
        '400':
          description: Data is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - tasks
      operationId: delete_all_tasks_for_user
      parameters:
      - name: X-API-AUTH-TOKEN
        in: header
        description: Auth token
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDeleteAllTasksForUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/rzd/tasks/{task_id}:
    delete:
      tags:
      - tasks
      operationId: delete_task_by_id_for_user
      parameters:
      - name: task_id
        in: path
        description: Task id
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-AUTH-TOKEN
        in: header
        description: Auth token
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDeleteTaskByIdForUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found for user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/login:
    post:
      tags:
      - users
      operationId: login
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginData'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseLogin'
        '400':
          description: Data is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/me:
    get:
      tags:
      - users
      operationId: me
      parameters:
      - name: X-API-AUTH-TOKEN
        in: header
        description: Auth token
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMe'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/signup:
    post:
      tags:
      - users
      operationId: signup
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpData'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSignUp'
        '400':
          description: Data is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/verify:
    get:
      tags:
      - users
      operationId: verify_user
      parameters:
      - name: verify_key
        in: query
        description: Verify token
        required: true
        schema:
          type: string
          format: uuid
      - name: redirect
        in: query
        description: Redirect link
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseLogin'
        '404':
          description: Verify token not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CreateTaskData:
      type: object
      required:
      - task_type
      - data
      properties:
        data:
          type: object
          additionalProperties:
            type: string
        task_type:
          type: string
    ErrorResponse:
      type: object
      required:
      - status
      - error
      properties:
        error:
          type: string
        status:
          type: string
    LoginData:
      type: object
      required:
      - username
      - password
      properties:
        password:
          type: string
        username:
          type: string
    ResponseCreateTask:
      type: object
      required:
      - status
      - data
      properties:
        data:
          $ref: '#/components/schemas/Task'
        status:
          type: string
    ResponseDeleteAllTasksForUser:
      type: object
      required:
      - status
      - data
      properties:
        data:
          type: string
        status:
          type: string
    ResponseDeleteTaskByIdForUser:
      type: object
      required:
      - status
      - data
      properties:
        data:
          type: string
        status:
          type: string
    ResponseListTasks:
      type: object
      required:
      - status
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        status:
          type: string
    ResponseLogin:
      type: object
      required:
      - status
      - data
      properties:
        data:
          type: string
        status:
          type: string
    ResponseMe:
      type: object
      required:
      - status
      - data
      properties:
        data:
          $ref: '#/components/schemas/UserReturn'
        status:
          type: string
    ResponseSignUp:
      type: object
      required:
      - status
      - data
      properties:
        data:
          $ref: '#/components/schemas/UserReturn'
        status:
          type: string
    SignUpData:
      type: object
      required:
      - username
      - email
      - password
      - repeat_password
      properties:
        email:
          type: string
        password:
          type: string
        repeat_password:
          type: string
        username:
          type: string
    Task:
      type: object
      required:
      - id
      - created_at
      - type_
      - data
      - user_id
      properties:
        created_at:
          type: string
          format: date-time
        data: {}
        id:
          $ref: '#/components/schemas/Thing'
        type_:
          type: string
        user_id:
          $ref: '#/components/schemas/Thing'
    UserReturn:
      type: object
      required:
      - id
      - username
      - is_verified
      - email
      - role
      - password
      properties:
        email:
          type: string
        id:
          $ref: '#/components/schemas/Thing'
        is_verified:
          type: boolean
        password:
          type: string
        role:
          type: string
        username:
          type: string